requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
sqlalchemy>=2.0
psycopg
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7.4
httpx>=0.27
python-dotenv>=1.0
redis>=5.0
python-multipart>=0.0.9
orjson>=3.9
